fix: refactor project
