+++ 2_2d_animation/src/collision.hpp
@@ -1,4 +1,5 @@
 #pragma once
+
 #include <glm/glm.hpp>
 #include <variant>
 #include <concepts>
